+++ src/main.js
@@ -4,19 +4,30 @@ require("prototype.tower");
 
 module.exports.loop = function() {
   "use strict";
-  
-  // Run creeps logic
+
+  // Make sure we have a Board
+  if (!Memory.board) Memory.board = {}
+
+  // Make sure we have a Queue
+  if (!Memory.queue) Memory.queue = [];
+
+  /**
+    Run Creeps
+    */
   for (let name in Game.creeps) {
-    // console.log(Game.creeps[name], Game.creeps[name].memory.role);
     Game.creeps[name].logic();
   }
 
-  // Run spawns logic
+  /**
+    Run Spawns
+    */
   for (let name in Game.spawns) {
     Game.spawns[name].logic();
   }
 
-  // Run towers logic
+  /**
+    Run Towers
+    */
   const towers = _.filter(Game.structures, s => s.structureType == STRUCTURE_TOWER);
   for (let tower of towers) {
     tower.logic();